+++ internal/table/retry.go
@@ -82,6 +82,7 @@ func (c *Client) retryOptions(opts ...table.Option) *table.Options {
 		RetryOptions: []retry.Option{
 			retry.WithTrace(c.config.TraceRetry()),
 			retry.WithBudget(c.config.RetryBudget()),
+			retry.WithIdempotent(c.config.DefaultIdempotent()),
 		},
 	}
 	for _, opt := range opts {